+++ yuna/destinations/__init__.py
@@ -1,4 +1,4 @@
 import logging
-logging.basicConfig(level=logging.DEBUG)
+
 logger = logging.getLogger("destinations")
 from . import mysql