+++ yuna/sources/__init__.py
@@ -1,4 +1,4 @@
 import logging
-logging.basicConfig(level=logging.DEBUG)
+
 logger = logging.getLogger("sources")
 from . import aliyun, windpy